Set restored fetch `this` bind to `undefined`
In strict mode (and modules), global functions are bound to `undefined`, not `globalThis`. The global fetch already has no `this` references when setting it in `STORE`, however still bind it to `undefined` in case there is a custom global fetch. On restoration, the bind is needed to match default binding to `undefined` instead of `globalThis`.