Clarify `intervalId` type
`ReturnType<typeof setInterval>` is needed because the return type of `setInterval` changes depending on runtime (e.g., web vs. Node.js). However, there is no impact to the behavior of the code regardless of return type.

Using `undefined` instead of `null` is useful to align with the default "do nothing" input that `clearInterval` takes, even if semantically incorrect.