Add TypeScript declaration file (.d.ts) generation
This is the cleanest way to make a JS library importable to a TS project. This approach is similar to Svelte's approach.

Note the dev dependency on TS is only used for .d.ts generation. The source code itself still does not use TS (or any dependencies other than Node.js) at all.